siam_1712022_3

+++ API/Dtos/InvestmentDetailDto.cs
@@ -13,7 +13,7 @@ namespace API.Dtos
         public DateTimeOffset FromDate { get; set; }
         public DateTimeOffset ToDate { get; set; }
         public int TotalMonth { get; set; }
-        public long ProposedAmount { get; set; }
+        public double ProposedAmount { get; set; }
         public string Purpose { get; set; }
         public int InvestmentInitId { get; set; }
