+++ Core/Entities/InvestmentRec.cs
@@ -17,7 +17,7 @@ namespace Core.Entities
         public DateTimeOffset FromDate { get; set; }
         public DateTimeOffset ToDate { get; set; }
         public int TotalMonth { get; set; }
-        public long ProposedAmount { get; set; }
+        public double ProposedAmount { get; set; }
         public string Purpose { get; set; }
         public bool? CompletionStatus { get; set; }
         public int Priority { get; set; }
